refactor(auth): 로그인·회원가입 페이지 Server/Client 컴포넌트 경계 분리
- components/auth/LoginForm.tsx를 신설해 로그인 페이지의 입력 상태·useAuth·useRouter·제출 처리를 Client Component로 분리
- components/auth/SignupForm.tsx를 신설해 회원가입 페이지도 동일하게 분리
- app/(host)/login/page.tsx, app/(host)/signup/page.tsx에서 'use client'를 제거하고 Server Component로 유지, 각각 LoginForm/SignupForm만 렌더링하도록 축소
- PR #112 코드래빗 리뷰 지적을 반영해 CLAUDE.md의 SSR/CSR 경계 규칙("공개 페이지는 SSR/SSG, 대시보드는 CSR")에 맞춤
- 이 커밋은 페어 모드가 아니라 Claude가 직접 작성함(사용자 코드 확인 필요 구간)

Refs #114
